reorganize and cleanup tests_tf folder
