+++ tests_tf/test_attacks.py
@@ -1,4 +1,7 @@
-from __future__ import absolute_import, division, print_function
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
 
 import unittest
 import numpy as np
