+++ tests_tf/test_utils.py
@@ -1,11 +1,12 @@
-from __future__ import absolute_import, division, print_function
-
-import unittest
+from __future__ import absolute_import
+from __future__ import division
+from __future__ import print_function
+from __future__ import unicode_literals
 
 import numpy as np
+import unittest
 
 from cleverhans import utils
-from cleverhans.utils_tf import kl_with_logits, l2_batch_normalize
 
 
 def numpy_kl_with_logits(q_logits, p_logits):
@@ -33,27 +34,6 @@ class TestUtils(unittest.TestCase):
         res = utils.other_classes(5, 2)
         res_expected = [0, 1, 3, 4]
         self.assertTrue(res == res_expected)
-
-    def test_l2_batch_normalize(self):
-        import tensorflow as tf
-        with tf.Session() as sess:
-            x = tf.random_normal((100, 1000))
-            x_norm = sess.run(l2_batch_normalize(x))
-            self.assertTrue(
-                np.allclose(np.sum(x_norm**2, axis=1), 1, atol=1e-6))
-
-    def test_kl_with_logits(self):
-        import tensorflow as tf
-        q_logits = tf.placeholder(tf.float32, shape=(100, 20))
-        p_logits = tf.placeholder(tf.float32, shape=(100, 20))
-        q_logits_np = np.random.normal(0, 10, size=(100, 20))
-        p_logits_np = np.random.normal(0, 10, size=(100, 20))
-        with tf.Session() as sess:
-            kl_div_tf = sess.run(kl_with_logits(q_logits, p_logits),
-                                 feed_dict={q_logits: q_logits_np,
-                                            p_logits: p_logits_np})
-        kl_div_ref = numpy_kl_with_logits(q_logits_np, p_logits_np)
-        self.assertTrue(np.allclose(kl_div_ref, kl_div_tf))
 
 
 if __name__ == '__main__':
